Re-arranging util package files.
